feat: add InternalGeneralFixedItems component
