+++ packages/apps/stock/src/components/organisms/index.js
@@ -43,3 +43,4 @@ export {default as ProductCardStockIndicatorList} from './ProductCardStockIndica
 export {default as StockCorrectionHeader} from './StockCorrectionHeader/StockCorrectionHeader';
 export {default as StockCorrectionBody} from './StockCorrectionBody/StockCorrectionBody';
 export {default as StockCorrectionFixedItems} from './StockCorrectionFixedItems/StockCorrectionFixedItems';
+export {default as InternalGeneralFixedItems} from './InternalGeneralFixedItems/InternalGeneralFixedItems';
